+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
+## Updated
+- Updated `FeedbackOptions` component to use real data [#203]
+
 ## Chore
 - Updated `sanitize-html` to `v2.17.3` and `dompurify` to `v3.4.0` due to security issues. Also, updated `prettier` to `v3.8.3` and `@dmptool/types` to `v3.1.4`, and `lodash` override to `v.4.18.1`. Removed overrides for `minimatch` and `test-exclude`.